+++ kafka-websocket-consumer/src/main/resources/avro/sale-report-v1.avsc
@@ -32,7 +32,7 @@
                 "type": "int"
             },
             {   "name": "sales_revenue",
-                "type": "double"
+                "type": "string"
             }
           ]
         }
