+++ kafka-websocket-producer/src/main/java/com/sonvu/producer/serializer/AvroSerializer.java
@@ -1,4 +1,4 @@
-package com.sonvu.producer.config;
+package com.sonvu.producer.serializer;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.avro.generic.GenericDatumWriter;
