+++ kafka-websocket-producer/src/main/resources/application-dev.properties
@@ -1,4 +1,4 @@
-app.interval-time=10000
+app.interval-time=20000
 app.data-file=/Users/macbook/Documents/Docs/pro_intellij/apache_kafka/deman-planing-dev-kafka/producer/data
 server.port=8080
 spring.kafka.producer.bootstrap-servers=websocket:9092
